Prime numbers example.
Now this example is pretty cool.
We have children that serve prime numbers and return them.
We have a supervisor that takes care of children and restarts them if
one fails.
We are using a usize as request type, but a fully fledged Response type
as a result.
info!() logs the successes.
Summary displays how long the program took, and how many failures there
were.